Add docs, unit tests, animations; fix build
Build fixes:
- Makefile: remove -I$(INCDIR) from CFLAGS (caused bare paths as linker args)
- Add missing libs: -lpspgu -lpspgum -lpsprtc
- socket.c: use fd_set/FD_ZERO/FD_SET instead of non-existent SceNetInetFdSet
- socket.c: use htons() from netinet/in.h instead of sceNetHtons
- wifi.c: fix strncpy truncation warning
- log.c: fix %u/%lu format mismatch for u32

Features:
- Needle animation with per-frame lerp (factor 0.15 in Racing Tuner, instant in OEM)
- GaugeDef.anim_val persists across frames via static local in dashboard_render_analog
- README.md: hardware, PIDs, build/deploy steps, controls, project structure
- USAGE.md: full setup guide, mode screenshots, troubleshooting table
- tests/: 64 native unit tests (no PSP required, gcc only)
  Covers: PID decode, hex parser, moving-avg filter, unit conversions,
          DTC formatter, reconnect backoff, ELM327 response roundtrip

+++ src/utils/log.c
@@ -22,7 +22,7 @@ void log_write(int level, const char *fmt, ...) {
     sceRtcGetCurrentTick(&tick);
     u32 ms = (u32)(tick / (sceRtcGetTickResolution() / 1000));
 
-    fprintf(g_log_fp, "[%08u][%s] ", ms, level_str[level]);
+    fprintf(g_log_fp, "[%08lu][%s] ", (unsigned long)ms, level_str[level]);
 
     va_list args;
     va_start(args, fmt);
